UPDATED role based access control, changed directory layout and added utility functions in utils.py
