+++ backend/utils/database.py
@@ -2,7 +2,9 @@ import os
 from pymongo import MongoClient
 
 # DB setup
-uri = os.getenv("MONGO_URI")
+# uri = os.getenv("MONGO_URI")
+uri = "mongodb+srv://domino_user:domino_user@road.idtag.mongodb.net/?retryWrites=true&w=majority&appName=ROAD"
+
 client = MongoClient(uri)
 db = client.road_project
 collection_users = db["users"]
