+++ backend/utils/utils.py
@@ -1,0 +1,86 @@
+"""
+author: Dominik Cedro
+date: 28.01.2025
+description: Util functions
+"""
+import bson
+import jwt
+from bson import ObjectId
+from fastapi import HTTPException
+from icecream import ic
+from jwt import InvalidTokenError
+from starlette import status
+
+from backend.models.models import UserResponse
+from backend.utils.database import collection_users
+from backend.utils.security import SECRET_KEY, ALGORITHM
+
+def extract_user_id_from_token(authorization: str) -> str:
+    """
+        Util function to retrieve user_id from jwt
+
+        Args:
+            authorization (str):  token
+
+        Returns:
+            user_id (str): id of user
+
+        Raises:
+            HTTP Exception 401 if user id is not existent or invalid token.
+     """
+    token = authorization.split(" ")[1]
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        user_id: str = payload.get("user_id")
+        ic("Extracted user_id from token:", user_id)
+        if user_id is None:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Could not validate credentials",
+                headers={"WWW-Authenticate": "Bearer"},
+            )
+        return user_id
+    except InvalidTokenError:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+
+
+def get_user_from_token(authorization: str):
+    token = authorization.split(" ")[1]
+    user_id = extract_user_id_from_token(token)
+    user = get_user_by_id(collection_users, user_id)
+    if not user:
+        raise HTTPException(status_code=404, detail="User not found")
+    return user
+
+
+def get_user_by_id(collection, user_id: str):
+    """
+        Retrieve user from db based on user_id - BSON
+
+        Args:
+            collection (Collection): The database collection to query.
+            user_id (str): The id of the user.
+
+        Returns:
+            UserResponse: full information about user excluding pesel, hashed password
+
+        Raises:
+            HTTP Exception 400 when invalid ID format
+            HTTP Exception 404 when user not found
+        """
+    ic("user id here is")
+    ic(user_id)
+    try:
+        user_dict = collection.find_one({"_id": ObjectId(user_id)})
+    except bson.errors.InvalidId:
+        ic("Invalid user ID format")
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid user ID format")
+    if user_dict:
+        user_dict["_id"] = str(user_dict["_id"])
+        return UserResponse(**user_dict)
+    else:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
